Refactor tests to mock `fs` and `glob`, and adjust imports for `cleanupTranslations`.
